Rename compModels to bodyCompModels
